requirements.txt
streamlit>=1.35.0
supabase>=2.5.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.18.0
anthropic>=0.25.0
python-docx>=1.1.0
PyMuPDF>=1.23.0
python-pptx>=0.6.23
Pillow>=10.0.0
